+++ app/src/main/java/com/rsschool/quiz/ui/pager/PagerContract.kt
@@ -1,7 +1,7 @@
 package com.rsschool.quiz.ui.pager
 
 import androidx.fragment.app.Fragment
-import com.rsschool.quiz.ui.utils.OnCurrentFragmentListener
+import com.rsschool.quiz.utils.OnCurrentFragmentListener
 
 interface PagerContract {
