+++ app/src/main/java/com/rsschool/quiz/utils/OnResultPageButtonsClickListener.kt
@@ -1,4 +1,4 @@
-package com.rsschool.quiz.ui.utils
+package com.rsschool.quiz.utils
 
 interface OnResultPageButtonsClickListener {
